// src/internal/types/messaging/messaging-types.ts
export type VoiceMessageJson = {
    /**
     * The message identifier.
     */
    voicemailId: string;

    /**
     * The party who has deposit this message.
     */
    from?: PartyInfoJson;

    /**
     * The message duration.
     */
    duration: number;

    /**
     * The date this message has been deposit.
     */
    date: Date;

    /**
     * Whether this message has been read.
     */
    unread: boolean;

    /**
     * Whether this message has high priority.
     */
    highPriority: boolean;
};